refactor: refactor pubkey / signing into separate keys.ts module
moved public key and signing utility functions to a dedicated module. added intellisense for new
apis.
